Support for ndk r10c
Android ndk r10c introduces a new warning that breaks fb-adb when we use android-21.

To fix this, this patch migrates from ptsname() to ptsname_r() when building
the Android binaries. Unfortunately, OS X does not provide ptsname_r(), so an
exception is required for this specific case.

Also, when compiling using Android-21 or later, the resulting binaries are
incompatible with older (<5.0) versions of Android. A warning has been added to
the README file to assist users.