Make bash(1) dependency explicit
On BSDs /bin is reserved for base utilities before /usr is mounted.
bash doesn't come with system by default but can be installed later
which is usually under either /usr/local/bin or /usr/pkg/bin.

zsh has a compatibility mode for various shells, including bash. It's
good enough to run the scripts here, so try as a fallback.