Always get a PTY even with newer adb
New versions of ADB have improved the shell command so that it's
clean, propagates error codes, and allocates PTYs unconditionally,
much like ssh.  This last change breaks fb-adb, since we run the
inferior adb with pipes for stdio; ADB thinks it's being run
non-interactively, skips allocating a PTY on the device side.
The remote shell doesn't have a pty itself, so _it_ thinks it's being
run non-interactively, so it doesn't print a prompt.  Meanwhile,
fb-adb hangs because it's waiting forever for a prompt.

This change just gets us back to using a PTY unconditionally.
We could take advantage of adb fixes to get rid of some of our
encoding logic, but that's a separate project.